test : 서비스 분리에 따른 testcode 분리
